Avoid generating unecessary accessors
Using a :CONC-NAME of NIL, we can avoid the generation of structure
accessors for the derived node types in the WBTREE package. This is a
lot cleaner than the formerly used |Don't use this..KEY| (and similar)
names.

The drawback is, that the structure slots have to be named slightly more
ugly, but that's ok, since those names should stay in the module and are
not exposed.