Rename HASHTRIE::NODE -> HASHTRIE:HASHTRIE
The top-level node definition structure is now a publicly accessibly
type for the benefit of code, which wants to dispatch on the "hashtrieness"
of objects. This is similar to what the "WBTREE" module already does
with its semi-private node structure.

Note, that the internals (in particular, the accessor for the nodes'
fields) are intentionally not disclosed, they remain a private detail
of the implementation.