Add WBTREE-MODIFY as alternative to WBTREE-UPDATE
The new function WBTREE-MODIFY behaves exactly like WBTREE-UPDATE
unless the key is already present. If the key is present, the new
value is determined by invoking the modifier function on the proposed
value and the one currently assigned. The modifier function may now
compute an effective new value, and can also control via secondary
return value, whether an update should be performed or not.

Hence, WBTREE-MODIFY is a generalization of WBTREE-UPDATE (every
operation possible via WBTREE-UPDATE is also possible using
WBTREE-MODIFY but not vice versa.)